fixV3

+++ SESMAN.Application/Services/BaseService.cs
@@ -48,6 +48,13 @@ namespace SESMAN.Application.Services
 
         public async Task<bool> DeleteAsync(Guid id)
         {
+            // veritabanı kısmında bu ID var mı kontrolü
+            var entity =  await _repository.GetByIdAsync(id);
+            if (entity == null)
+            {
+                return false;
+            }
+
             await _repository.DeleteAsync(id);
             return true;
         }
